+++ src/cpaldevice.rs
@@ -51,7 +51,8 @@ pub struct CpalCaptureDevice {
     pub devname: String,
     pub host: CpalHost,
     pub samplerate: usize,
-    pub resampler_conf: config::Resampler,
+    pub resampler_type: config::Resampler,
+    pub resampler_profile: config::ResamplerProfile,
     pub enable_resampling: bool,
     pub capture_samplerate: usize,
     pub chunksize: usize,
@@ -458,8 +459,9 @@ impl CaptureDevice for CpalCaptureDevice {
         let bits_per_sample = self.sample_format.bits_per_sample() as i32;
         let sample_format = self.sample_format.clone();
         let enable_resampling = self.enable_resampling;
-        let resampler_conf = self.resampler_conf.clone();
-        let async_src = resampler_is_async(&resampler_conf);
+        let resampler_type = self.resampler_type.clone();
+        let resampler_profile = self.resampler_profile.clone();
+        let async_src = resampler_is_async(&resampler_type);
         let silence_timeout = self.silence_timeout;
         let silence_threshold = self.silence_threshold;
         let stop_on_rate_change = self.stop_on_rate_change;
@@ -470,7 +472,8 @@ impl CaptureDevice for CpalCaptureDevice {
                 let mut resampler = if enable_resampling {
                     debug!("Creating resampler");
                     get_resampler(
-                        &resampler_conf,
+                        &resampler_type,
+                        &resampler_profile,
                         channels,
                         samplerate,
                         capture_samplerate,